Create MainBot.py
My Python Oanda Bot 
This is a program that trades based on EMA crosses. For the best signals, we're using the 5, 10, 20, 50, 100 day EMAs. * NOTE * This program does NOT WORK, it's incomplete.. I am posting this here for EDUCATIONAL USE ONLY. I am not responsible for any of your losses because of my work. You can use this code at your own risk.
The summary of the program is that we're tracking the 5, 10, 20, 50, and 100 day EMAs. We make trade decisions based on the proximity of the short term (5, 10, 20) day EMAs with the longer term ones. Crosses are signals to "reverse" the position, while proximity is a signal to reverse. 